requirements.txt
# Core Dependencies
ib_async>=2.0.1
pandas>=1.5.0
numpy>=1.21.0
pydantic>=2.0.0
python-dotenv>=0.19.0

# Database
sqlalchemy>=2.0.0
alembic>=1.8.0

# Cache
redis>=4.3.0

# Display and CLI
rich>=13.0.0
click>=8.1.0

# Async
asyncio>=3.4.3
aiohttp>=3.8.0

# Machine Learning
scikit-learn>=1.1.0
tensorflow>=2.10.0
opencv-python>=4.6.0
joblib>=1.2.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Technical Analysis
scipy>=1.9.0
ta-lib>=0.4.0  # Optional, requires system library

# API and Web
requests>=2.28.0
websocket-client>=1.3.0

# Sentiment Analysis
textblob>=0.17.0
nltk>=3.7.0
praw>=7.6.0  # Reddit API
tweepy>=4.10.0  # Twitter API

# Configuration
PyYAML>=6.0

# Time and Date
pytz>=2022.1
python-dateutil>=2.8.0

# Utilities
tqdm>=4.64.0
colorama>=0.4.5

# Development and Testing
pytest>=7.1.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
pytest-mock>=3.7.0
black>=22.0.0
ruff>=0.0.261
mypy>=0.961
pre-commit>=2.19.0

# Type stubs
types-PyYAML
types-requests
types-redis
pandas-stubs

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
myst-parser>=0.18.0

# Monitoring and Logging
structlog>=22.0.0
sentry-sdk>=1.9.0  # Optional error tracking

# Performance
ujson>=5.4.0  # Faster JSON
orjson>=3.7.0  # Even faster JSON
msgpack>=1.0.4  # Binary serialization

# Additional ML/Data Science
xgboost>=1.6.0
lightgbm>=3.3.0
optuna>=3.0.0  # Hyperparameter optimization
shap>=0.41.0  # Model explainability

# Financial Data
yfinance>=0.1.74  # Yahoo Finance fallback
finnhub-python>=2.4.0  # Alternative data source
alpaca-py>=0.8.0  # Alternative broker